requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
Django==4.0.5
django-environ==0.8.1
django-model-utils==4.2.0
djangorestframework==3.13.1
environ==1.0
mysqlclient==2.1.0
pytz==2022.1
sqlparse==0.4.2
